change int to double

+++ src/integrate/ensemble_ti.cuh
@@ -53,7 +53,7 @@ protected:
   double lambda = 0;
   double pe, espring;
   // spring constants
-  std::map<std::string, int> spring_map;
+  std::map<std::string, double> spring_map;
   GPU_Vector<double> gpu_k;
   std::vector<double> cpu_k;
   GPU_Vector<double> gpu_espring;
